Exploratory Data Analytics in Agriculture
The climate and local weather conditions have a key influence upon agriculture. Weather events such as heat, drought, heavy rain or frost have always governed the everyday life of farmers. There is probably no industry that is as closely linked to the weather as agriculturt. These data are collected by weather stations, The results of this data analysis provide significant insights to farmers and agronomists about how to improve the production, minimise the farming operational costs, manage risks, and protect the environment.